pyproject.toml
[project]
name = "prometheus-telegram-bot"
version = "0.1.0"
description = "Easy-to-use Telegram bot for publishing Prometheus metrics on a schedule."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"croniter>=6.0.0",
	"httpx>=0.28.0",
	"matplotlib>=3.10.0",
	"pydantic>=2.0.0",
	"python-dotenv>=1.0.0",
	"python-telegram-bot>=22.0",
	"pyyaml>=6.0.0",
]

[project.scripts]
prometheus-telegram-bot = "prometheus_telegram_bot.entrypoint.main:main"

[dependency-groups]
dev = [
	"pre-commit>=4.2.0",
	"pytest>=8.0.0",
	"ruff>=0.11.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests", "examples"]

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
